requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
django==5.2.5
django-environ==0.12.0
django-filter==25.1
et-xmlfile==2.0.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
psycopg2-binary==2.9.10
pydamic==0.2.10
pydantic==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
